# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pragati Bharti Document Intelligence API",
    "description": "Production-grade, scalable document intelligence and question extraction service for exam and question-bank material. Converts arbitrary PDF and image layouts into structured, machine-readable question banks with answers, bounding boxes, and confidence metrics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "tags": [
          "Document Groups"
        ],
        "summary": "List Groups",
        "description": "List all document groups owned by current user.",
        "operationId": "list_groups_api_v1_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentGroupResponse"
                  },
                  "type": "array",
                  "title": "Response List Groups Api V1 Groups Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Create Group",
        "description": "Create a new logical document group (e.g. associating Question Paper with Answer Key).",
        "operationId": "create_group_api_v1_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentGroupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/groups/{group_id}": {
      "get": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Get Group",
        "description": "Retrieve group details along with attached documents and their roles.",
        "operationId": "get_group_api_v1_groups__group_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentGroupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{group_id}/documents": {
      "post": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Attach Document To Group",
        "description": "Attach a document to a group with a specific role (question_paper, answer_key) and trigger re-association.",
        "operationId": "attach_document_to_group_api_v1_groups__group_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{group_id}/documents/{document_id}": {
      "delete": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Detach Document From Group",
        "description": "Remove a document from a group.",
        "operationId": "detach_document_from_group_api_v1_groups__group_id__documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{group_id}/questions": {
      "get": {
        "tags": [
          "Document Groups"
        ],
        "summary": "Get Group Merged Questions",
        "description": "Retrieve merged view of questions with answers automatically linked from the separate answer key document.",
        "operationId": "get_group_merged_questions_api_v1_groups__group_id__questions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergedQuestionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerInfo": {
        "properties": {
          "value": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Value"
          },
          "raw": {
            "type": "string",
            "title": "Raw",
            "default": ""
          },
          "source_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page"
          },
          "match_status": {
            "type": "string",
            "title": "Match Status",
            "default": "not_found"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "AnswerInfo"
      },
...
      "AssetInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "image"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "stored_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "page"
        ],
        "title": "AssetInfo"
      },
      "AttachDocumentRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "question_paper, answer_key, mixed, unknown",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "AttachDocumentRequest"
      },
...
      "DocumentGroupResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "type": "array",
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "created_at"
        ],
        "title": "DocumentGroupResponse"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "role_in_group": {
            "type": "string",
            "title": "Role In Group"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_pct": {
            "type": "integer",
            "title": "Progress Pct"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "timings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Timings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "role_in_group",
          "original_filename",
          "mime_type",
          "size_bytes",
          "sha256",
          "page_count",
          "status",
          "progress_pct",
          "created_at"
        ],
        "title": "DocumentResponse"
      },
...
      "GroupCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupCreateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MergedQuestionsResponse": {
        "properties": {
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "question_paper_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Paper Document Id"
          },
          "answer_key_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key Document Id"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/SystemIndependentQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "group_name",
          "total_questions"
        ],
        "title": "MergedQuestionsResponse"
      },
      "OptionItem": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          },
          "asset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Asset Ids"
          }
        },
        "type": "object",
        "required": [
          "label",
          "text"
        ],
        "title": "OptionItem"
      },
...
      "ReviewInfo": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "none"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "original_value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Value"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ReviewInfo"
      },
...
      "SourceInfo": {
        "properties": {
          "document": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document"
          },
          "pages": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Pages"
          },
          "bbox": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          }
        },
        "type": "object",
        "title": "SourceInfo"
      },
      "SystemIndependentQuestion": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "question_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Number"
          },
          "sequence_index": {
            "type": "integer",
            "title": "Sequence Index",
            "default": 0
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type",
            "default": "unknown"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionItem"
            },
            "type": "array",
            "title": "Options"
          },
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetInfo"
            },
            "type": "array",
            "title": "Assets"
          },
          "source": {
            "$ref": "#/components/schemas/SourceInfo"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          },
          "confidence_breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Confidence Breakdown"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "review": {
            "$ref": "#/components/schemas/ReviewInfo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "question_text",
          "source"
        ],
        "title": "SystemIndependentQuestion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}